+++ src/contacts/qcontact.h
@@ -137,7 +137,7 @@ private:
     QSharedDataPointer<QContactData> d;
 };
 
-Q_CONTACTS_EXPORT uint qHash(const QContact& key);
+Q_CONTACTS_EXPORT size_t qHash(const QContact& key);
 #ifndef QT_NO_DEBUG_STREAM
 Q_CONTACTS_EXPORT QDebug operator<<(QDebug dbg, const QContact& contact);
 #endif
