+++ src/contacts/qcontactcollection.h
@@ -89,12 +89,12 @@ public:
     QVariant extendedMetaData(const QString &key) const;
 
 private:
-    friend Q_CONTACTS_EXPORT uint qHash(const QContactCollection &key);
+    friend Q_CONTACTS_EXPORT size_t qHash(const QContactCollection &key);
     friend class QContactManagerEngine;
     QSharedDataPointer<QContactCollectionData> d;
 };
 
-Q_CONTACTS_EXPORT uint qHash(const QContactCollection &key);
+Q_CONTACTS_EXPORT size_t qHash(const QContactCollection &key);
 
 #ifndef QT_NO_DEBUG_STREAM
 Q_CONTACTS_EXPORT QDebug operator<<(QDebug dbg, const QContactCollection &collection);
