+++ src/contacts/qcontactdetail.h
@@ -200,7 +200,7 @@ private:
 #endif
 };
 
-Q_CONTACTS_EXPORT uint qHash(const QContactDetail& key);
+Q_CONTACTS_EXPORT size_t qHash(const QContactDetail& key);
 #ifndef QT_NO_DEBUG_STREAM
 Q_CONTACTS_EXPORT QDebug operator<<(QDebug dbg, const QContactDetail& detail);
 #endif
