+++ src/organizer/qorganizercollection.h
@@ -88,12 +88,12 @@ public:
     QVariant extendedMetaData(const QString &key) const;
 
 private:
-    friend Q_ORGANIZER_EXPORT uint qHash(const QOrganizerCollection &key);
+    friend Q_ORGANIZER_EXPORT size_t qHash(const QOrganizerCollection &key);
     friend class QOrganizerManagerEngine;
     QSharedDataPointer<QOrganizerCollectionData> d;
 };
 
-Q_ORGANIZER_EXPORT uint qHash(const QOrganizerCollection &key);
+Q_ORGANIZER_EXPORT size_t qHash(const QOrganizerCollection &key);
 
 #ifndef QT_NO_DEBUG_STREAM
 Q_ORGANIZER_EXPORT QDebug operator<<(QDebug dbg, const QOrganizerCollection &collection);
