+++ src/organizer/qorganizercollectionid.h
@@ -82,7 +82,7 @@ private:
 inline bool operator<(const QOrganizerCollectionId &id1, const QOrganizerCollectionId &id2)
 { return id1.managerUri() != id2.managerUri() ? id1.managerUri() < id2.managerUri() : id1.localId() < id2.localId(); }
 
-inline uint qHash(const QOrganizerCollectionId &id)
+inline size_t qHash(const QOrganizerCollectionId &id)
 { return qHash(id.localId()); }
 
 #ifndef QT_NO_DATASTREAM
