+++ src/organizer/qorganizeritemid.cpp
@@ -109,7 +109,7 @@ QT_BEGIN_NAMESPACE_ORGANIZER
 */
 
 /*!
-    \fn uint qHash(const QOrganizerItemId &id)
+    \fn size_t qHash(const QOrganizerItemId &id)
     \relates QOrganizerItemId
 
     Returns the hash value for \a id.
