+++ src/organizer/qorganizerrecurrencerule.h
@@ -131,13 +131,13 @@ private:
 };
 
 //hash functions
-Q_ORGANIZER_EXPORT uint qHash(const QOrganizerRecurrenceRule &rule);
+Q_ORGANIZER_EXPORT size_t qHash(const QOrganizerRecurrenceRule &rule);
 
 #ifndef QT_NO_DEBUG_STREAM
 Q_ORGANIZER_EXPORT QDebug operator<<(QDebug dbg, const QOrganizerRecurrenceRule &rule);
 #endif // QT_NO_DEBUG_STREAM
 
-inline uint qHash(QOrganizerRecurrenceRule::Month month)
+inline size_t qHash(QOrganizerRecurrenceRule::Month month)
 {
     return static_cast<uint>(month);
 }
